+++ project/main.gd
@@ -21,6 +21,11 @@ func _ready() -> void:
 	_init_level_choice_popup()
 	_update_user_info()
 
+	# Run basic Android tests
+	if OS.get_name() == "Android" and ResourceLoader.exists("res://test/test_on_android.gd"):
+		var tests = load("res://test/test_on_android.gd").new()
+		tests.run_tests()
+
 
 func _init_level_choice_popup() -> void:
 	var popup: PopupMenu = level_choice.get_popup()