+++ src/sentry/native/native_sdk.cpp
@@ -1,6 +1,7 @@
 #include "native_sdk.h"
 
 #include "sentry.h"
+#include "sentry/common_defs.h"
 #include "sentry/contexts.h"
 #include "sentry/level.h"
 #include "sentry/native/native_event.h"
@@ -22,9 +23,6 @@
 #ifdef DEBUG_ENABLED
 #include <godot_cpp/classes/time.hpp>
 #endif
-
-#define _SCREENSHOT_FN "screenshot.jpg"
-#define _VIEW_HIERARCHY_FN "view-hierarchy.json"
 
 namespace {
 
@@ -70,7 +68,7 @@ void _save_screenshot(const Ref<SentryEvent> &p_event) {
 	}
 	last_screenshot_frame = current_frame;
 
-	String screenshot_path = "user://" _SCREENSHOT_FN;
+	String screenshot_path = "user://" SENTRY_SCREENSHOT_FN;
 	DirAccess::remove_absolute(screenshot_path);
 
 	if (!DisplayServer::get_singleton() || DisplayServer::get_singleton()->get_name() == "headless") {
@@ -117,7 +115,7 @@ inline void _save_view_hierarchy() {
 	uint64_t start = Time::get_singleton()->get_ticks_usec();
 #endif
 
-	String path = "user://" _VIEW_HIERARCHY_FN;
+	String path = "user://" SENTRY_VIEW_HIERARCHY_FN;
 	DirAccess::remove_absolute(path);
 
 	if (OS::get_singleton()->get_thread_caller_id() != OS::get_singleton()->get_main_thread_id()) {
@@ -340,7 +338,7 @@ String NativeSDK::capture_event(const Ref<SentryEvent> &p_event) {
 	return _uuid_as_string(uuid);
 }
 
-void NativeSDK::initialize() {
+void NativeSDK::initialize(const PackedStringArray &p_global_attachments) {
 	ERR_FAIL_NULL(OS::get_singleton());
 	ERR_FAIL_NULL(ProjectSettings::get_singleton());
 
@@ -387,30 +385,13 @@ void NativeSDK::initialize() {
 		sentry_options_set_backend(options, NULL);
 	}
 
-	// Attach LOG file.
-	// TODO: Decide whether log-file must be trimmed before send.
-	if (SentryOptions::get_singleton()->is_attach_log_enabled()) {
-		String log_path = ProjectSettings::get_singleton()->get_setting("debug/file_logging/log_path");
-		if (FileAccess::file_exists(log_path)) {
-			log_path = log_path.replace("user://", OS::get_singleton()->get_user_data_dir() + "/");
-			sentry_options_add_attachment(options, log_path.utf8());
+	for (const String &path : p_global_attachments) {
+		sentry::util::print_debug("adding attachment \"", path, "\"");
+		if (path.ends_with(SENTRY_VIEW_HIERARCHY_FN)) {
+			sentry_options_add_view_hierarchy(options, path.utf8());
 		} else {
-			WARN_PRINT("Sentry: Log file not found. Make sure \"debug/file_logging/enable_file_logging\" is turned ON in the Project Settings.");
+			sentry_options_add_attachment(options, path.utf8());
 		}
-	}
-
-	// Attach screenshot.
-	if (SentryOptions::get_singleton()->is_attach_screenshot_enabled()) {
-		String path = OS::get_singleton()->get_user_data_dir().path_join(_SCREENSHOT_FN);
-		DirAccess::remove_absolute(path);
-		sentry_options_add_attachment(options, path.utf8());
-	}
-
-	// Attach view hierarchy (aka scene tree info).
-	if (SentryOptions::get_singleton()->is_attach_scene_tree_enabled()) {
-		String path = OS::get_singleton()->get_user_data_dir().path_join(_VIEW_HIERARCHY_FN);
-		DirAccess::remove_absolute(path);
-		sentry_options_add_view_hierarchy(options, path.utf8());
 	}
 
 	// Hooks.